Fix Utility::getScaleToFill/Fit potential integer division

+++ include/Misc/Utility.h
@@ -10,7 +10,7 @@ namespace Utility
     template<typename T>
     float getScaleToFill(const sf::Vector2<T>& initialDimensions, const sf::Vector2<T>& fillDimensions)
     {
-        return std::max(fillDimensions.x / initialDimensions.x, fillDimensions.y / initialDimensions.y);
+        return std::max(static_cast<double>(fillDimensions.x) / initialDimensions.x, static_cast<double>(fillDimensions.y) / initialDimensions.y);
     }
 
     /// Get scale to fit given dimensions (without overlapping) while preserving aspect ratio
